+++ python/sglang/srt/layers/moe/dwdp/transport.py
@@ -11,7 +11,12 @@ import torch
 import torch.distributed as dist
 from cuda.bindings import driver as cuda
 
-from sglang.srt.cuda_vmm_utils import (
+from sglang.srt.layers.moe.dwdp.layout import (
+    DwdpExpertLayout,
+    LayerWeightSpecs,
+    MnnvlHandleSet,
+)
+from sglang.srt.utils.cuda_vmm_utils import (
     VmmReservation,
     align_down,
     align_up,
@@ -22,11 +27,6 @@ from sglang.srt.cuda_vmm_utils import (
     import_peer_handle,
     make_device_allocation_prop,
     tensor_from_pointer,
-)
-from sglang.srt.layers.moe.dwdp.layout import (
-    DwdpExpertLayout,
-    LayerWeightSpecs,
-    MnnvlHandleSet,
 )
 
 logger = logging.getLogger(__name__)
